refactor: extract workout persistence contract
